Make a gene model plot for a chromosomal location

Uses the Gviz module to show transcripts and exon locations for a given gene id.

Usage

geneModelPlot(ensembl_species, chromosome, start, end)

Arguments

ensembl_species

Ensembl species definition like 'mmuscululus'

chromosome

Chromosome number

start

Chromosome start coordinate

end

Chromosome end coordinate