Make a numeric field with selectable associated cardinality (>, < ..).

A wrapper around numericInput, providing an inline label and an associated field specifying how the value should be applied, i.e. 'greater than this value' etc.

Usage

cardinalNumericField(
  id,
  cardinal_id,
  label,
  value,
  cardinality = "<=",
  step = NA,
  min = NA,
  max = NA,
  tooltip = NULL
)

Arguments

id

ID to use for the numeric field

cardinal_id

ID to use for the cardinality field

label

Label

value

Default value

cardinality

Default cardinality

step

Passed to numericInput

min

Passed to numericInput

max

Passed to numericInput

tooltip

Optional tooltip text explaining the field, shown via a help icon next to the label (see withHelpIcon)

Value

out An HTML tag object that can be rendered as HTML using as.character()